refactor: 경기 결과 view 분리
