+++ src/main/java/racing_car_with_winner/controller/GameController.java
@@ -5,8 +5,9 @@ import racing_car_with_winner.model.CarNameInput;
 import racing_car_with_winner.model.GameAttemptCount;
 import racing_car_with_winner.model.MovableStrategy;
 import racing_car_with_winner.model.car.Cars;
-import racing_car_with_winner.view.GameDashBoard;
+import racing_car_with_winner.view.MoveResultView;
 import racing_car_with_winner.view.UserOutputInterface;
+import racing_car_with_winner.view.WinnerView;
 
 public class GameController {
     private final MovableStrategy movableStrategy;
@@ -19,15 +20,22 @@ public class GameController {
 
     public void play(CarNameInput carNameInput, GameAttemptCount gameAttemptCount) {
         Cars cars = new Cars(new CarFactory(carNameInput).build());
-
         GameAttemptCount playCount = gameAttemptCount;
 
+
+        userOutputInterface.show("실행결과\n");
         do {
             playCount = playCount.play();
+
             cars = cars.moveCars(movableStrategy);
-            cars.getCars().forEach(car -> new GameDashBoard(userOutputInterface).showMoveResult(car.getCarName(), car.getMoveResult()));
+
+            cars.getCars().stream()
+                .map(car -> new MoveResultView(car.getCarName(), car.getMoveResult()).getMessage())
+                .forEach(userOutputInterface::show);
+
+            userOutputInterface.show("\n");
         } while(!playCount.isEnded());
 
-        new GameDashBoard(userOutputInterface).showWinner(cars.findWinner());
+        userOutputInterface.show(new WinnerView(cars.findWinner()).getMessage());
     }
 }
