+++ src/main/java/racing_car_with_winner/view/ConsoleUserInputInterface.java
@@ -7,7 +7,7 @@ public class ConsoleUserInputInterface implements UserInputInterface {
 
     @Override
     public void show(String message) {
-        System.out.println(message);
+        System.out.print(message);
     }
 
     @Override
