+++ src/main/java/racing_car_with_winner/view/ConsoleUserOutputInterface.java
@@ -3,6 +3,6 @@ package racing_car_with_winner.view;
 public class ConsoleUserOutputInterface implements UserOutputInterface {
     @Override
     public void show(String message) {
-        System.out.println(message);
+        System.out.print(message);
     }
 }
